resume_screening_ai/requirements.txt
streamlit
flask
requests
pandas
numpy
scikit-learn
nltk
spacy
transformers
pillow
PyMuPDF
python-docx
